This repo contains the task 2 and additional challenge
To enhance your learning experience and add more functionality to your project, we encourage you to take on the following challenges:

Add a feature to delete notes within the application.
Use IndexedDB for storing notes instead of local storage. For easier interaction with IndexedDB, consider using the Dexie.js library.